[ARM]: Driver for the global timer
This is a simple driver for the global timer module found in the
Cortex A9-MP cores from revision r1p0 onwards. This should be able
to perform the functions of the system timer and the local timer
in an SMP system, however so far it has only been tested in a
non SMP environment. If used SMP a mechanism will need to be added
to ensure the clock event is not instantiated twice for the initial
core.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>