stm: FDMA support for STiH415, Kernel Config
Kernel Config option added from the original patch.
It looks like this got dropped during a merge somewhere

This is a first attempt to support the FDMA hardware on STiH415.
Currently only FDMA0 (in the MPE) is configured, although data is
provided for all FDMA and cross bar devices. Currently the FDMA
code won't cope if more than one crossbar is configured.

Note that as the FDMA support relies on the SH DMA API, this has been
ported across to maintain the API.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>
Signed-off-by: Kieran Bingham <kieran.bingham@st.com>