fdma: Support multiple crossbars
This patch adds support to the FDMA driver for multiple request line
crossbars. Each crossbar is given an integer ID in its platform data,
and each FDMA specifies which crossbar it is attached to in /its/
platform data. fdma_req_allocate() will then use the correct crossbar
for the channel in use.

A crossbar ID of "-1" (which all our SOCs currently have), is defined as
"there is only one crossbar", which is then designated as crossbar #0.

Signed-off-by: Chris Smith <chris.smith@st.com>