stm_mali: Fix the makefile for out of tree builds
EXTRA_CFLAGS is deprecated and ccflags-y should be used

In addition, having a ' ' (space) between -I and <path> is an error and
compilation fails during out of tree builds (using O= or KBUILD_OUTPUT=)

We also have to make use of $(srctree) to get access to the full path
where needed for out of tree builds.

We should use := instead of = for speed.

Signed-off-by: Andre Draszik <andre.draszik@st.com>
Signed-off-by: Stephen Gallimore <stephen.gallimore@st.com>