stm: stih415: Add the IDs of first and last FDMA using a FDMA XBAR
This patch allows a FDMA XBAR to store the first and last IDs of the FDMAs
connected to it. We then use the stored IDs to calculate the correct offset to
program.

This fixes an issue on STiH415 where there are five FDMAs and two XBARs. The
first XBAR is connected to FDMAs 0, 1, and 2. The second XBAR is connected to
FDMAs 3 and 4. When programming the second XBAR we incorrectly use the FDMA ID
(3) to calculate the register to write. We should be using 0. Thus storing the
first ID of the FDMAs using the XBAR we can subtract this from the FDMA ID
(i.e. 3 - 3) and calcute the correct register offset.

Signed-off-by: John Boddie <john.boddie@st.com>