b2000: review Kconfig and support B2048A eMMC/MMC modules (v3)
This patch reviews the B2000 Kconfig for
SH4 and ARM, fixes some bug in the board's setup
when use the B2035 and adds the support for
the B2048 boards. There are two versions of
these boards. A board is with the eMMC and
another one is with the MMC slot.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Signed-off-by: Youssef TRIKI <youssef.triki@st.com>