b2000: Fix coherent dma and speculation issue
By default, the PL310 L2 cache changes shared normal memory uncached
accesses into cacheable no-allocate ones. In other words, it will do
a cache lookup in L2 rather than simply bypassing straight to L3 (DDR
in this case.

The problem with this behaviour is that if the processor speculates
through the kernel alias of the dma page, it will pull the line into
the L2 cache. In this case, the processor will see the stale value
when looking through the dma page, as it will hit the L2 cache rather
than looking at DDR. This has been observed to happen with the stmmac
driver and the EHCI driver.

The solution is to disable this behaviour. The stale speculative entry
in the L2 cache will do no harm as it will always be bypassed when
accessed via the dma alias, and there is no legitimate reason to access
it via the kernel alias. When the page is freed and reallocated it will
be zeroed so there is no problem with the stale value suddenly appearing
in a new page.

Signed-off-by: David McKay <david.mckay@st.com>