snd_stm: Correct STiH415 uniperipheral number of samples to read calculation
There is a bug in the uniperipheral hardware (RnDHV00031683) where the
memory fully read interrupt is generated after receiving the number of
samples programmed in UNIPERIF_I2S_FMT register plus one for 16/16 format
and plus two for 16/0 format. Under Havana this results in an underflow
every 50s or so and is documented in Bugzilla 13979.

This patch takes the solution identified in Bugzilla 13979 and applies it
to the kernel tree.

Signed-off-by: Daniel Thompson <daniel.thompson@st.com>
Signed-off-by: John Boddie <john.boddie@st.com>
Signed-off-by: Chris Smith <chris.smith@st.com>