pcie_stm: Add support for PCI Express switches
Previously the config read/write functions did not know how to generate
type 1 config accesses. These accesses are required to access devices
that are on the other side of a PCIe switch.

There are also a couple of other issues solved here that came out as a
result of adding the type 1 access.

* There appears to be a race between writing to the FUNC0_BDF_NUM and
  doing the read that actual issues the config cycle. This can sometimes
  result in the wrong device being read. This is solved by a readback.

* The wrapper logic ignores the status code returned by the PCIe core
  when doing configuration accesses. Thus the configuration cannot
  tell the difference between UR (unsupported request) codes and CRS
  (configuration retry) requests. Heuristics are used to retry the
  request, but this means that there will be delays while probing links
  that do not have devices connected to them.

One issue that it is not possible to work around here is that the
wrapper logic does not export INTB, INTC, INTD. This means that legacy
interrupts will not work from behind a switch as the bridge will
"swizzle" the interrupts, transmuting INTA->INTB etc.

Signed-off-by: David McKay <david.mckay@st.com>