stm: Driver for AMBA to STBUS convertor
In virtually all ST SOCS there is IP such as SATA, USB etc that comes
With an AMBA interface. There is a common piece of hardware that
converts between STBUS and AMBA. This code introduces common code to
manage this rather than each individual driver managing it.

The API is quite simple, there is a data structure that describes how
the bridge should be set up. There is then a create() function that sets
things up, and then an init() function that actually sets the bridge up.
The driver that uses this API is responsible for providing an ioremaped
pointer to the actual hardware block itself. There is no release method,
because all memory allocation is done via the devres api, so when the
calling device is removed  this will be automatically be removed as
well.

There are two different versions of the IP, type1 is the older version
that is gradually being replaced by the more capable type2.

There is no power management hardware on this block, however after
resume you must re-initialise the convertor.

This patch also provides a debugfs entry under the amba-stbus-bridge
directory. Each device will have a file and the values of the registers
are displayed as hex and fully decoded.

Signed-off-by: David McKay <david.mckay@st.com>