b2000: fix cn22 selection and remove warnings
This patch fixes a problem when select the B2048 that erroneously
enabled the GMAC0.

The patch also removes some warnings when compile.
In the end, the B2032A GMII modified Card on GMAC0 is now considered as
a subcase of the B2032A when configure the Kernel.

Patch is for both arm and sh.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>