snd_stm: Correct reference counting for mmap
The stm sound core code calls snd_pcm_mmap_data_open() in the mmap
method. The effect of this is to increment the reference count for
the buffer. The problem is that this is also incremented by the
vm_ops, so the effect is that the internal checks on the reference
count fail because it never goes back to zero.

The mystery is why this worked and was thought to be required on
the earlier kernels, should be investigated.

Signed-off-by: David McKay <david.mckay@st.com>