stih415: Add workaround to GMAC0 SMI bit 15 Bug(v2).
On GMAC0, it is noticed that some mdio register reads return value with
bit 15 set, however the waveforms on board show that they are
transmitted correctly. Which is a BUG at SOC level. Each board behaved
differently w.r.t which registers returned with bit 15 set.  This issue
is very much noticable when used with Gbit Switch or Gbit Mode.

Example:
On IP1001 PHY
with STMMAC-MDIO mii-bus driver:
register 2 reads are returned as 0x8243, however actual value is 0x0243.

with mdio-gpio mii-bus driver:
register 2 reads are returned as 0x0243, which is CORRECT.

Using mdio-gpio driver, the all register reads return expected values,
which is a workaround till the bug is fixed at SOC level.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>