stm_nand_afm: Use correct endian define
Previously the code was using __LITTLE_ENDIAN__ to select the correct
code for little endian. However this symbol is actually sh specific
and is not defined for the arm, so the wrong code was selected.

This patch replace it with __LITTLE_ENDIAN which is defined for all
architectures (see asm/byteorder.h)

Signed-off-by: David Mckay <david.mckay@st.com>