mtd_nand: 4bit on die ECC correct dev_id usage
When the 4bit ECC support was added for the 2.6.32 kernel the dev_id
was passed in as an integer. In later kernels this change to a int *,
so this makes the corresponding change in the 4bit ECC code.

change was made in  d1e1f4e42b5df063712ca2926e50c07b95c96b96

Signed-off-by: David McKay <david.mckay@st.com>