b2000: Add support for NAND via AFM and FLEX
Adds support for NAND for the stih415 and b2000 board. This uses
either the AFM or FLEX controller, EMI is currently unsupported.

At the moment you must have the stih415 mode pins jumpered
to use the Hamming controller rather than the BCH controller. If you
do not then the code will lock up when it tries to talk to the
controller. This will change in the not too distant future as the driver
will select the correct mode dynamically.

Signed-off-by: David McKay <david.mckay@.st.com>