stm_nand_emi: Update emi driver for 3.x kernel
There have been some minor changes in the way partitions are handled in
the MTD code, most notably that CONFIG_MTD_PARTITIONS doesn't exist
anymore. This patch makes the necessary changes to update the emi
driver.

This driver is currently untested as the stih415 doesn't support the EMI
bit banging driver yet (if it ever will).

Signed-off-by: David McKay <david.mckay@st.com>