stm_nand: Introduce flex/afm specific memcpy_toio
The FLEX and AFM drivers both use memcpy_toio() and memcpy_fromio()
routines for various reasons, mainly in the case of AFM to shovel
commands at the sequencer hardware.

The controller can only be targeted by 32 bit writes, byte transactions
will cause an imprecise exception error in the ARM. The problem is that
the ARM memcpy_toio() routines are totally unoptimised and simply do a
byte loop. This causes the exception to be generated.

On the SH4 memcpy_toio() is fully optimised and will always write as
efficiently as possible, using some cache optimisations so it is worth
using it.

This patch introduces some code to do the right thing on the ARM. I
chose to implement it as a new function rather than redefine
memcpy_toio() to avoid confusion and indicate that there is something
special going on.

Signed-off-by: David McKay <david.mckay@st.com>