sh_stm: Re-introduce ST PMB implementation
Unfortunately the ST and Renesas implementations of the PMB
support are very different. They assume different starting
conditions at boot, and use different algorithms to allocate
PMBs. Ideally we would like to combine the two implementations
into one, but this requires quite a bit of work.

This patch introduces the ST implementation from the 2.6.32 kernel
into the 3.x tree. It is not active, this patch is in effect a
rename. There might be a way to do this in git directly.

Later patches will in effect bolt the Renesas interface onto the
ST code, so that either implementation can be used

Signed-off-by: David McKay <david.mckay@st.com>