sh: Correct VMALLOC_START address
The sh mainline still does not have a dedicated allocator for
consistent memory allocations. This is required as the consistent
allocator can be called from an interrupt context. Quite why the
Renasis parts escape this is unknown.

When the merge from the ST tree was done VMALLOC_START was set back to
the start of the P3SEG, rather than from CONSISTENT_BASE. This meant
that the vmalloc address space and the address space used for the
consistent allocator overlap with unfortunate results.

Signed-off-by: David McKay <david.mckay@st.com>