ARM: Re-introduce dmac_clean/dmac_invalidate
The dmac_clean_range() and dmac_inv_range() functions were dropped in
later kernels. This patch re-introduces them as they are used by Havana
internally. It defines them in terms of the dmac_map/unmap functions
which will haev the same effect without having to reintroduce the actual
functions themselves.

A better short term solution is to add in the ioremap_flush() family for
ARM again. The longer terms solution is to change to the DMA API
instead, but this gives problems with bpa2 at the moment.

Signed-off-by: David McKay <david.mckay@st.com>