sh: Fix consistent alloc faults
The fault handler was not handling faults for the consistent alloc
area, which corresponds to P3SEG -> VMALLOC_START. This bug was masked
for some time because user processes inherit the kernels page table
which will most likely contain this mapping. This means that the primary
fault handler will handle this case.

Signed-off-by: David McKay <david.mckay@st.com>