sh: Use P4 addresses directly in ioremap()
Due to the SH4's segmented architecture devices that are mapped
physically in the P4 address range can be accessed directly by
using the actual physical address. There is no need to remap these
addresses if we are accessing them uncached, so we just use them
directly.

A further advantage of this is that it allows early drivers to call
ioremap() before the pmbs are set up, provided of course they live in
the P4 address space.

Signed-off-by: David McKay <david.mckay@st.com>