pci_stm: Update ST PCI support
There have been changes in the PCI infrastructure that mean that the PCI
infrastructure needs updating. The first change is that there is no
support for IO routines in the machine vector anymore.  The code now
assumes that you have sensible hardware and can initiate an IO
transaction on the bus by simply deferencing memory. Unfortunately of
course our hardware doesn't work like this, as you actually have to run
code. For the moment, I have chosen to simply disable IO entirely, it
will just blow up if you try and use it. Further down the line we can
either re-introduce the IO framework, or attempt to use the existing
infrastructure and point the IO region at some special region and
intercept the faults to do the right thing.

The other problem is that the generic sh pci code is now checking the
return values from resource_claim(). Since the PCI memory window was in
the platform devicei resources, this fails and bombs out because the
platform infrastructure will add the memory resources for us. The best
solution to this problem is to remove the memory window from the
platform resources, and pass in the addresses via the platform data
instead.

This patch updates both PCI and PCI express.

Signed-off-by: David McKay <david.mckay@st.com>