sh_stm: Bolt in ST PMB implementation
This patch introduces the changes needed to the ST PMB
implementation to make it interchangeable with the Renesas
one. This is a simple restruction to supprot the bolt()
API, and to change where pmb_init() is called from.

It is now possible to compile the kernel with either ST or Renesas code
and the kernel will boot with either. For the Renesas code, you must
boot the kernel with the se29=1 target pack parameter, as the Renesas
code requires that two PMBs are set up at boot, one cached and one
uncached.  The ST implementation only requires that there is one
uncached mapping available.

Signed-off-by: David McKay <david.mckay@st.com>