stm_mmc: add a new own structure to manage SDHCI pltfm d.d.
As already done on several ARM platform, special callback or
flags can be passed  through a specific structure for our ARCH
(i.e. stm_mmc_platform_data). This provides, for example, the
flag to pass from the board if there is the EMMC or call
the PAD manager to configure our resources.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>