stm:snd Workaround SMP deadlocks in the ALSA clock management
This patch fixes an issue debugging spinlocks as detailed in Bugzilla 16113.

In the ALSA code we register a clock wrapper with the generic STM clock
framework to allow fine grained adjustment of the audio player clock rate.

When the audio player calls into the STM clock framework to enable a clock or
set the clock rate, it will call into the ALSA wrapper. The ALSA wrapper does
what it needs to do, then ultimately ends up calling back into the STM clock
framework to do the actual enable or set the clock rate.

The problem is that on entry to the STM clock framework, an access spinlock is
locked. When ALSA wrapper then calls back into the STM clock framework, it
tries to lock the spinlock again. When the option SPINLOCK_DEBUG is enabled in
the kernel build configuration this recursive lock is detected as a bug.

There are two main solutions:
 1. Call the ALSA clock wrapper directly from the audio players.
 2. Call the generic clock framework internal functions directly.

For simplicity this patch goes with option 2. It has the smallest change foot
print, but at the cost of importing even more knowledge of the STM clock
framework into the ALSA wrapper code. I think this is a reasonable trade off
as the ALSA wrapper code already has some knowledge of the STM clock framework.

Signed-off-by: Daniel Thompson <daniel.thompson@st.com>
Signed-off-by: John Boddie <john.boddie@st.com>