stm:snd:fli7610 Add uniperipheral player support
This patch adds support for the uniperipheral players on FLi7610 and has
been tested on the FLi76xxHDK01 board. Note that due to low yields of this
SoC, you may not have all uniperipheral players working!

For SPDIF output to work on the HDK board, the SPDIF recepticle must be
removed, rotated 180 degrees and re-soldered to the board.

No audio will be heard from the LS/AUX/HP connections unless you route the
audio through the AATV.

This patch does not provide support for the uniperipheral readers!

Signed-off-by: John Boddie <john.boddie@st.com>