libata: Set read/write flag in ata_pio_sector()
The SFF data transfer functions take a parameter to indicate
if we are doing a READ (0) or a WRITE (1) . However, the ata_pio_sector()
function is actually passing the masked ATA flags value through.
So if you are doing a write you will get ATA_TFLAG_WRITE which is
0x8 rather than WRITE (1).

Most drivers still function correctly because they do something like:

if(rw) {
  /* Write */
} else {
  /* Read */
}

OR

if (rw == READ) {
  /* Read */
} else {
  /* Write */
}

Which will still work correctly.

However, pata_legacy.c and the stmicro sata driver instead do something
like:

if (rw == WRITE) {
	/* Write */
} else {
	/* Read */
}

Which will fail.

Reported-by: Laurent Cans <laurent.cans@sagemcom.com>
Signed-off-by: David McKay <david.mckay@st.com>