sh_stm: Add export.h for ilc3,stm-l2-cache,bpa2
We include export.h to get the definitions for EXPORT_SYMBOL. Since
these files cannot be built as modules, there is not need for the full
module.h so we use  export.h which gives a cut down version of the full
module support. See commit f50169324df4ad942e544386d136216c8617636a for
more details.

Signed-off-by: David McKay <david.mckay@st.com>