stm_b2000: Use ARM_PATCH_PHYS_VIRT in defconfig
This option allows us to produce a kernel that will boot regardless
of where it is located in physical memory. It does this by dynamically
patching the code at runtime so that virt_to_phys() etc work correctly.
This is useful for us during the transition to contiguous LMI, where
the load address will shift from 0x40000000 to 0x60000000. A kernel
built with this option will run at either address.

This option cannot be used if the kernel is signed in any way, since
it patches code dynamically. Thus it cannot realistically be used in
a secure system.

Signed-off-by: David McKay <david.mckay@st.com>