stm:snd Increase FDMA trigger limit for uniperipheral player/reader
Previously the uniperipheral was not taking full advantage of the FDMA by
having the FDMA trigger limit set to a maximum of 120-bytes. This patch
rectifies this by setting the FDMA trigger limit to the maximum 160-bytes.

This patch also corrects the calculation of the number of bytes to transfer
as performed in function snd_stm_pcm_transfer_bytes(). By changing the
check condition to 'transfer_bytes * 2 <= max_transfer_bytes', We now allow
up to the full max transfer of bytes.

Signed-off-by: John Boddie <john.boddie@st.com>