stm:fdma FDMA changes to manipulate node pointers
This change is a nasty hack, necessitated out of hardware bugs in Orly and
is part of the fix to Bugzilla 14269.

In summary, there is a problem with the Orly hardware - the HDMI is not
clocked when uniperipheral player 0 is in stop mode. The solution to keep
the HDMI clocked is to never stop uniperipheral player 0. To achieve this
when playback ends we switch to playing back a silent parking buffer. Then
when playback resumes we switch from the parking buffer to the playback
buffer.

The easy solution would be to stop playback as normal, then restart with
playback from the parking buffer. Unfortunately this is not suitable as it
takes too long to stop and restart.

The actual solution used is to fiddle with the FDMA next node pointers!
When the playback buffer is to be stopped, the FDMA should automatically
switch to the parking buffer for us. Then when playback buffer resumes the
FDMA should automatically switch back to the playback buffer.

To achieve this, we need some new API provided by this patch that will
allow us to manipulate the FDMA next pointer whilst running:
- dma_set_next_node()
  Set the FDMA next node whilst FDMA is running.
- dma_get_node_ptr()
  Get the FDMA current and next node pointers from the DMA compiled list.
- dma_configure_node_params()
  Update the FDMA channel parameters whilst running

Signed-off-by: sanjay jha <sanjay@sanjay.jha@st.com>
Signed-off-by: John Boddie <john.boddie@st.com>