stm:snd Add parking buffer for Bugzilla 14269
This change is a nasty hack, necessitated out of hardware bugs in Orly and
is part of the fix to Bugzilla 14269.

In summary, there is a problem with the Orly hardware - the HDMI is not
clocked when uniperipheral player 0 is in stop mode. The solution to keep
the HDMI clocked is to never stop uniperipheral player 0. To achieve this
when playback ends we switch to playing back a silent parking buffer. Then
when playback resumes we switch from the parking buffer to the playback
buffer.

The easy solution would be to stop playback as normal, then restart with
playback from the parking buffer. Unfortunately this is not suitable as it
takes too long to stop and restart.

The actual solution used is to fiddle with the FDMA next node pointers!
When the playback buffer is to be stopped, the FDMA should automatically
switch to the parking buffer for us. Then when playback buffer resumes the
FDMA should automatically switch back to the playback buffer.

The main changes to the uniperipheral player driver:
- Parking is optional and must be enabled for each uniperipheral player.
- Parking mode always plays linear pcm silence.
- Playback period buffer is split over multiple FDMA nodes to allow faster
  transition from playback to parking buffer as we don't have to wait for
  whole period to end.
- A state machine manages the transitions from playback to parking.
- FDMA callbacks are used throughout to actually transition from playback
  to parking buffers and vice versa.
- FDMA callbacks are used to update periods elapsed. This is no longer
  interrupt based.
- When parking is enabled, we only change the hardware parameters when the
  new parameters to set are different.

Signed-off-by: sanjay jha <sanjay@sanjay.jha@st.com>
Signed-off-by: John Boddie <john.boddie@st.com>