stm:snd Correct programming of uniperipheral player channel status bits
This patch fixes Bugzilla 15274 by moving the programming of the
uniperipheral player channel status bits into a separate function.

The driver will configure a default channel status assuming linear pcm
mode with an unspecified sampling frequency. This should be enough to hear
sound from a connected TV (certainly works with a Samsung and Toshiba TV).

It is assumed that the user will specify the correct channel status during
audio output. However, if no sampling rate is specified in the channel
status, then the current sampling rate will be set, and the correct sample
max word size will be set to 24-bit for linear pcm mode, and 16-bit for
encoded (compressed) mode.

Signed-off-by: John Boddie <john.boddie@st.com>