stm:snd Update uniperipheral player channel status whilst running
This patch fixes Bugzilla 16164. It will allow the uniperipheral player
channel status to be updated whilst the player is running. It will also
ensure that when the operating mode if changed from pcm to encoded
(compressed) mode and back, that the validity bits for the left and right
sub-frames are updated accordingly.

It is worth noting that when in encoded mode Player2 does not required the
additional encoded mode hardware features. Instead, Player2 provides
correctly formatted data and expects the uniperipheral player to operate in
normal pcm mode, but with the channel status and validity bits set for
encoded mode.

Signed-off-by: John Boddie <john.boddie@st.com>