kptrace: Update to latest version
This patch contains the full changeset to bring KPTrace fully up to
date. It includes the following headline changes:

 - Refactoring to support multiple architectures
 - ARM support
 - SMP support
 - Kernel 3.x support

It also includes the changes proposed during the last patch review of
this code. It was agreed at that review that these changes should be
submitted as a single large patch.

Tested on STiH415/SMP, STiH415/UP and STi7105.

Signed-off-by: Chris Smith <chris.smith@st.com>