pci_stm: Change irq mapping function for PCI
This is another patch in preparation for PCIe on ARM. The problem here
is that the superh PCI layer expects the BSP to provide a global
function called pcibios_map_platform_irq() to map pci slots to irq
numbers. The ARM PCI layer passes in a function for each pci bus rather
than one global function. So what we do here is pass in the function as
part of the pci data, and the pci glue code will call this. This means
that we can drop the stm_pci_legacy_irq(), as the pci glue code can
access the legacy interrupt directly.

Since at the moment there is no PCI (not express!) hardware present on
any of the ST ARM chips, this change may not be needed. However it means
the glue layer can be written to handle both PCI and PCIe cleanly which
is worth doing anyway.

Signed-off-by: David McKay <david.mckay@st.com>