pcie_stm: Add PCI glue logic for ARM
Introduces the glue logic to interface the ARM PCI subsystem to the
STM PCIe driver. The main PCIe driver needs to work on both superh and
ARM, and unfortunately the way the PCI subsystems work is not common
across different arches. It is a real shame that there is no common
framework for all to use, but it would be a lot of work.

This patch  also increase the number of interrupts available to allow
some space for MSI interrupts. We allocate 32 which is more than enough.

Signed-off-by: David McKay <david.mckay@st.com>