stm_stih415: Add PCI Express on Orly
Adds in support for the PCIe cell on the Orly for the b2000 board.
This plugs in the Orly specific values into the PCI framework.

Many b2000 boards have the option resistor for the PCIe clock populated
wrongly. R130 should be in 2-3 and R131 in 1-2. This will give a 100MHz
clock. If these resistors are not right then PCIe will not work.

Signed-off-by: David McKay <david.mckay@st.com>