pcie_stm: Change link up polling interval to 12ms
This increases the time interval at which we check if the link is up
from 8ms to 12ms. The reason for this is that the specification states
that the time interval from detect.quiet -> detect.active is 12ms. Also
the units were previously in microseconds, which was causing warnings
because the number was too large.

Signed-off-by: David McKay <david.mckay@st.com>