pcie_stm: Pass in LMI memory address/size
This is preparatory work for adding ARM support to the PCIe driver. At
the moment, we are using CONFIG_MEMORY_START and CONFIG_MEMORY_SIZE to
find out where main memory is. The PCIe driver needs to know this so it
can open up the window so the PCIe devices can bus master onto main
memory. Unfortunately these symbols are SH specific, so instead the
parameters are passed in along with the other memory information.

Signed-off-by: David McKay <david.mckay@st.com>