pcie_stm: Make ARM specific interrupt changes for MSI
The ARM has several pecularities to do with interrupts. The first is
that every interrupt has got to be marked as valid with
set_irq_flags(irq, IRQF_VALID). Quite what the point of this is I don't
know.

The second difference is the the GIC interrupts are of the fasteoi type.
So once the multiplexor interrupt has been fired, the eoi function for
the interrupt has to be called or no more interrupts will be triggered.
The ARM provides a function called chained_irq_exit() which will do
this.

This patch adds these functions into the MSI code.

IMHO these functions need not be ARM specific, they should be part of
the generic framework.

Signed-off-by: David McKay <david.mckay@st.com>