arm:dma-mapping: Remove spurious DMA API debug errors doing range based DMA syncs with non zero offsets
The arm arch specific dma-mapping versions of dma_sync_single_range_for_device
and dma_sync_single_range_for_cpu call the non-range versions of
the debug_dma_xxx functions, with the mapped device address modified by the
"offset" parameter specified as part of the range being synced.

This results in the DMA debug functions incorrectly reporting that the DMA
mapping being synced was never allocated, because the mapping address has been
modified, if the "offset" parameter is not zero.

The versions of the DMA functions in asm-generic use the "range" versions of the
DMA debug functions, this patch modifies the ARM code to do the same.

Signed-off-by: Stephen Gallimore <stephen.gallimore@st.com>