stm:snd Fix uniperipheral player state machine crash
This patch fixes an issue in the uniperipheral player state machine where
if you suspend the program playing audio over a uniperipheral player, then
resume the program, a crash happens.

The issue is due to the state machine logic not correctly handling all the
required states. This patch adds in the code to support the missing states.

This patch also splits the parking mode into it's three parts; init, loop,
and exit. This allows more precise state machine handling and better error
recovery.

The uniperipheral player will also attempt to detect when the FDMA is
processing the parking loop or exit nodes in order to prevent any queries
to the FDMA for updating the audio pointer (that could result in a crash).

Signed-off-by: John Boddie <john.boddie@st.com>