stm: fdma: Use extension API for functions which manipulate node pointers
Commit 3fd1abe98e4ba3f87de11b5fe99296f51b753bd7
    stm:fdma FDMA changes to manipulate node pointers
added a number of new functions to manipulate FDMA node pointers.
However these were added by extending the SH DMA API in the drivers/stm
directory, which is a copy of the SH DMA API in arch/sh/drivers/dma,
without updating the original.

By converting these new functions use use the extension API, all these
changes to the SH DMA API can be removed, and it should now be compatible
with both the original code (used on SH) and the copy (used on ARM).

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>