stm:fli7610: Allow CLKA_PLL0LS clock to set enable.
ClockgenA has 2 PLL's PLL_sys_0 and PLL_sys_1 and are respectively of
type PLL_PG_1600x and PLL_PG_800x. However PLL_sys_o has two outputs one
at 900 and other at 450. Existing code only allows 900 output to be
enabled, but when trying to enable 450 output it just doesn not pass the
check. Due this behaviour the stm clk framework disables the parent,
which is PLL_sys_0.

This patch is an fix to Newman VTAC problem.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>