stm:gpio:ARM: set EOI from gpio chained interrupt handler.
As Stih415 and Fli7610 uses GIC which has fasteoi setup for the primary
interrupt, and interrupt handler is expected to write to "End of
Interrupt Register" when its finished handling the interrupt, as part of
fasteoi flow control.

This patch fixes missing logic in stm-gpio driver specific to ARM
fasteoi flow control.

Orignially detected as part of stlinux bugzilla #17853 triage.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>