stm_gpio: Remove erroneous warning message
There is no way to manipulate the irq internal state in the 3.x kernels,
so there is no legitimate way to set the irq as pending again. Therefore
this is as good as we can get without introducing more generic code.
Hence we can delete the warning message and the code stub.

Signed-off-by: David McKay <david.mckay@st.com>