kptrace: Fix mutex tracing when lock debugging is enabled.
Tracing of mutex_init() was broken when CONFIG_DEBUG_LOCK_ALLOC ("Lock
debugging: detect incorrect freeing of live locks") was enabled, because
in this case the function mutex_init() doesn't exist.

This patch uses an aliased tracepoint on mutex_init_nested() to make a
mutex_init tracepoint available to userspace in both cases. This
requires the creation of an extra kptrace_create_*_tracepoint()
variant, kptrace_create_late_aliased_tracepoint(), as this is the first
time we've needed that particular combination of tracepoint properties.

Also renamed the existing create_late_tracepoint() as
kptrace_create_late_tracepoint() for consistency.

Signed-off-by: Chris Smith <chris.smith@st.com>