kptrace: Remove unnecessary strcmp from userspace_store_attrs
The sysfs function userspace_store_attrs allows userspace to write to
the KPTrace trace buffer by writing to the file "userspace/new_record".
"new_record" is the only attribute of "userspace", so there is no need
to strcmp the name every time, slowing things down.

Thanks to Marc Titinger for spotting this!

Signed-off-by: Chris Smith <chris.smith@st.com>