kptrace: Convert from a sysdev to a "real" device
The kernel no longer provides sysdevs so KPTrace, like all the other
sysdevs, has to be converted to a real device. This patch makes the
change.

This patch also trivially fixes a couple of warnings, and in doing so
flips the sense of the return value of create_sysfs_tree() to the right
way around.

Signed-off-by: Chris Smith <chris.smith@st.com>