sh_stm: Remove support for ST40RA based chips
The ST40RA is no longer supported. This was the very first SH4 chip
from STMicroelectronics, and has not been supported in reality for some
time. This patch puts the nail in the coffin by removing it completely.

Board support for the mb360 and mb374 are removed, along with hardware
unique to the ST40RA like the PCI cell.

We also remove the EPLD irq support, this was used by other chips but
they have also been removed so this is now completely dead.

Signed-off-by: David McKay <david.mckay@st.com>