stm_nand_bch: do not program reserved registers
Previously, we loaded the BCH sequence by programing a block of 16 registers.
However, two registers within the block are marked as 'reserved' in the NANDi
spec.  On ARM platforms, this results in an imprecise data exception.  This
patch skips the two reserved registers when programming the BCH sequence, which
is probably the correct approach for SH as well as ARM.

Signed-off-by: Angus Clark <angus.clark@st.com>